LPS-112607 portal-search-elasticsearch7-impl: Rename SidecarElasticsearchConnectionManager to SidecarManager and move to sidecar package
https://issues.liferay.com/browse/LPS-112607
