+++ modules/apps/portal-search-elasticsearch7/portal-search-elasticsearch7-impl/src/main/java/com/liferay/portal/search/elasticsearch7/internal/sidecar/Sidecar.java
@@ -577,13 +577,13 @@ public class Sidecar {
 				}
 			}
 
-			SidecarComponentUtil.disableSidecarElasticsearchConnectionManager();
+			SidecarComponentUtil.disableSidecarManager();
 
 			if (_log.isInfoEnabled()) {
 				_log.info("Restarting sidecar Elasticsearch process");
 			}
 
-			SidecarComponentUtil.enableSidecarElasticsearchConnectionManager();
+			SidecarComponentUtil.enableSidecarManager();
 		}
 
 	}
