+++ modules/apps/portal-search-elasticsearch7/portal-search-elasticsearch7-impl/src/main/java/com/liferay/portal/search/elasticsearch7/internal/sidecar/SidecarComponentUtil.java
@@ -14,8 +14,6 @@
 
 package com.liferay.portal.search.elasticsearch7.internal.sidecar;
 
-import com.liferay.portal.search.elasticsearch7.internal.connection.SidecarElasticsearchConnectionManager;
-
 import org.osgi.service.component.ComponentContext;
 import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
@@ -26,17 +24,15 @@ import org.osgi.service.component.annotations.Component;
 @Component(immediate = true, service = {})
 public class SidecarComponentUtil {
 
-	public static void disableSidecarElasticsearchConnectionManager() {
+	public static void disableSidecarManager() {
 		if (_componentContext != null) {
-			_componentContext.disableComponent(
-				SidecarElasticsearchConnectionManager.class.getName());
+			_componentContext.disableComponent(SidecarManager.class.getName());
 		}
 	}
 
-	public static void enableSidecarElasticsearchConnectionManager() {
+	public static void enableSidecarManager() {
 		if (_componentContext != null) {
-			_componentContext.enableComponent(
-				SidecarElasticsearchConnectionManager.class.getName());
+			_componentContext.enableComponent(SidecarManager.class.getName());
 		}
 	}
