+++ modules/apps/portal-search-elasticsearch7/portal-search-elasticsearch7-impl/src/main/java/com/liferay/portal/search/elasticsearch7/internal/sidecar/SidecarManager.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.search.elasticsearch7.internal.connection;
+package com.liferay.portal.search.elasticsearch7.internal.sidecar;
 
 import com.liferay.petra.process.ProcessExecutor;
 import com.liferay.petra.string.StringBundler;
@@ -25,8 +25,11 @@ import com.liferay.portal.search.elasticsearch7.internal.configuration.Elasticse
 import com.liferay.portal.search.elasticsearch7.internal.configuration.ElasticsearchConfigurationObserverComparator;
 import com.liferay.portal.search.elasticsearch7.internal.configuration.ElasticsearchConfigurationWrapper;
 import com.liferay.portal.search.elasticsearch7.internal.configuration.OperationModeResolver;
-import com.liferay.portal.search.elasticsearch7.internal.sidecar.ProcessExecutorPathsImpl;
-import com.liferay.portal.search.elasticsearch7.internal.sidecar.Sidecar;
+import com.liferay.portal.search.elasticsearch7.internal.connection.ConnectionConstants;
+import com.liferay.portal.search.elasticsearch7.internal.connection.ElasticsearchConnectionBuilder;
+import com.liferay.portal.search.elasticsearch7.internal.connection.ElasticsearchConnectionManager;
+import com.liferay.portal.search.elasticsearch7.internal.connection.ElasticsearchInstancePaths;
+import com.liferay.portal.search.elasticsearch7.internal.connection.ElasticsearchInstancePathsBuilder;
 import com.liferay.portal.search.elasticsearch7.settings.SettingsContributor;
 
 import java.nio.file.Files;
@@ -48,8 +51,7 @@ import org.osgi.service.component.annotations.ReferencePolicyOption;
  * @author Tina Tian
  */
 @Component(enabled = true, immediate = true, service = {})
-public class SidecarElasticsearchConnectionManager
-	implements ElasticsearchConfigurationObserver {
+public class SidecarManager implements ElasticsearchConfigurationObserver {
 
 	@Override
 	public int compareTo(
@@ -213,8 +215,7 @@ public class SidecarElasticsearchConnectionManager
 	@Reference
 	protected Props props;
 
-	private static final Log _log = LogFactoryUtil.getLog(
-		SidecarElasticsearchConnectionManager.class);
+	private static final Log _log = LogFactoryUtil.getLog(SidecarManager.class);
 
 	private final Set<SettingsContributor> _settingsContributors =
 		new ConcurrentSkipListSet<>();
